# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"CodeMentor API","description":"API service for CodeMentor website","version":"0.1.0"},"paths":{"/v1/healthcheck":{"get":{"tags":["healthcheck"],"summary":"Check service health","description":"Endpoint to check the health of the service.","operationId":"healthcheck_v1_healthcheck_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/challenges/":{"post":{"tags":["challenges"],"summary":"Create","operationId":"create_v1_challenges__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChallengeBody"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_Challenge_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["challenges"],"summary":"Get Challenges","operationId":"get_challenges_v1_challenges__get","parameters":[{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_list_ReadChallenge__"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/challenges/{challenge_id}":{"put":{"tags":["challenges"],"summary":"Update","operationId":"update_v1_challenges__challenge_id__put","parameters":[{"name":"challenge_id","in":"path","required":true,"schema":{"type":"integer","title":"Challenge Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChallengeBody"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_Challenge_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["challenges"],"summary":"Delete","operationId":"delete_v1_challenges__challenge_id__delete","parameters":[{"name":"challenge_id","in":"path","required":true,"schema":{"type":"integer","title":"Challenge Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_Challenge_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["challenges"],"summary":"Get Challenge","operationId":"get_challenge_v1_challenges__challenge_id__get","parameters":[{"name":"challenge_id","in":"path","required":true,"schema":{"type":"integer","title":"Challenge Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_ReadChallenge_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/challenges/{challenge_id}/tags":{"post":{"tags":["challenges"],"summary":"Add Tags","operationId":"add_tags_v1_challenges__challenge_id__tags_post","parameters":[{"name":"challenge_id","in":"path","required":true,"schema":{"type":"integer","title":"Challenge Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"integer"},"title":"Tags"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_ChallengeTagsLink_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/tags/create":{"post":{"tags":["tags"],"summary":"Create","operationId":"create_v1_tags_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TagBody"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_Tag_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/tags/update/{tag_id}":{"put":{"tags":["tags"],"summary":"Update","operationId":"update_v1_tags_update__tag_id__put","parameters":[{"name":"tag_id","in":"path","required":true,"schema":{"type":"integer","title":"Tag Id"}},{"name":"new_name","in":"query","required":true,"schema":{"type":"string","title":"New Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_Tag_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/tags/":{"get":{"tags":["tags"],"summary":"Get Tags","operationId":"get_tags_v1_tags__get","parameters":[{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_list_Tag__"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/tags/{tag_id}":{"get":{"tags":["tags"],"summary":"Get","operationId":"get_v1_tags__tag_id__get","parameters":[{"name":"tag_id","in":"path","required":true,"schema":{"type":"integer","title":"Tag Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_Tag_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/tags/search/":{"get":{"tags":["tags"],"summary":"Tags Search","operationId":"tags_search_v1_tags_search__get","parameters":[{"name":"tag_name","in":"query","required":true,"schema":{"type":"string","minLength":1,"title":"Tag Name"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_list_Tag__"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/tags/delete/{tag_id}":{"delete":{"tags":["tags"],"summary":"Delete","operationId":"delete_v1_tags_delete__tag_id__delete","parameters":[{"name":"tag_id","in":"path","required":true,"schema":{"type":"integer","title":"Tag Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_Tag_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/auth/github":{"get":{"tags":["auth"],"summary":"Github Login","operationId":"github_login_v1_auth_github_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/auth/github/callback":{"get":{"tags":["auth"],"summary":"Github Callback","operationId":"github_callback_v1_auth_github_callback_get","parameters":[{"name":"code","in":"query","required":true,"schema":{"type":"string","title":"Code"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/auth/me":{"get":{"tags":["auth"],"summary":"Me","operationId":"me_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_ReadUser_"}}}}},"security":[{"HTTPBearer":[]}]}},"/v1/challenges/{challenge_id}/solutions":{"post":{"tags":["solutions"],"summary":"Create Solution","operationId":"create_solution_v1_challenges__challenge_id__solutions_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"challenge_id","in":"path","required":true,"schema":{"type":"integer","title":"Challenge Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChallengeSolutionBody"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_ReadChallengeSolution_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["solutions"],"summary":"List Solutions","operationId":"list_solutions_v1_challenges__challenge_id__solutions_get","parameters":[{"name":"challenge_id","in":"path","required":true,"schema":{"type":"integer","title":"Challenge Id"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":50,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_list_ReadChallengeSolution__"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/solutions/{solution_id}":{"get":{"tags":["solutions"],"summary":"Get Solution","operationId":"get_solution_v1_solutions__solution_id__get","parameters":[{"name":"solution_id","in":"path","required":true,"schema":{"type":"integer","title":"Solution Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_ReadChallengeSolution_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["solutions"],"summary":"Delete Solution","operationId":"delete_solution_v1_solutions__solution_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"solution_id","in":"path","required":true,"schema":{"type":"integer","title":"Solution Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_NoneType_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"APIResponse_ChallengeTagsLink_":{"properties":{"success":{"type":"boolean","title":"Success"},"data":{"anyOf":[{"$ref":"#/components/schemas/ChallengeTagsLink"},{"type":"null"}]},"error":{"anyOf":[{"$ref":"#/components/schemas/ErrorDetail"},{"type":"null"}]}},"type":"object","required":["success"],"title":"APIResponse[ChallengeTagsLink]"},"APIResponse_Challenge_":{"properties":{"success":{"type":"boolean","title":"Success"},"data":{"anyOf":[{"$ref":"#/components/schemas/Challenge"},{"type":"null"}]},"error":{"anyOf":[{"$ref":"#/components/schemas/ErrorDetail"},{"type":"null"}]}},"type":"object","required":["success"],"title":"APIResponse[Challenge]"},"APIResponse_NoneType_":{"properties":{"success":{"type":"boolean","title":"Success"},"data":{"type":"null","title":"Data"},"error":{"anyOf":[{"$ref":"#/components/schemas/ErrorDetail"},{"type":"null"}]}},"type":"object","required":["success"],"title":"APIResponse[NoneType]"},"APIResponse_ReadChallengeSolution_":{"properties":{"success":{"type":"boolean","title":"Success"},"data":{"anyOf":[{"$ref":"#/components/schemas/ReadChallengeSolution"},{"type":"null"}]},"error":{"anyOf":[{"$ref":"#/components/schemas/ErrorDetail"},{"type":"null"}]}},"type":"object","required":["success"],"title":"APIResponse[ReadChallengeSolution]"},"APIResponse_ReadChallenge_":{"properties":{"success":{"type":"boolean","title":"Success"},"data":{"anyOf":[{"$ref":"#/components/schemas/ReadChallenge"},{"type":"null"}]},"error":{"anyOf":[{"$ref":"#/components/schemas/ErrorDetail"},{"type":"null"}]}},"type":"object","required":["success"],"title":"APIResponse[ReadChallenge]"},"APIResponse_ReadUser_":{"properties":{"success":{"type":"boolean","title":"Success"},"data":{"anyOf":[{"$ref":"#/components/schemas/ReadUser"},{"type":"null"}]},"error":{"anyOf":[{"$ref":"#/components/schemas/ErrorDetail"},{"type":"null"}]}},"type":"object","required":["success"],"title":"APIResponse[ReadUser]"},"APIResponse_Tag_":{"properties":{"success":{"type":"boolean","title":"Success"},"data":{"anyOf":[{"$ref":"#/components/schemas/Tag"},{"type":"null"}]},"error":{"anyOf":[{"$ref":"#/components/schemas/ErrorDetail"},{"type":"null"}]}},"type":"object","required":["success"],"title":"APIResponse[Tag]"},"APIResponse_list_ReadChallengeSolution__":{"properties":{"success":{"type":"boolean","title":"Success"},"data":{"anyOf":[{"items":{"$ref":"#/components/schemas/ReadChallengeSolution"},"type":"array"},{"type":"null"}],"title":"Data"},"error":{"anyOf":[{"$ref":"#/components/schemas/ErrorDetail"},{"type":"null"}]}},"type":"object","required":["success"],"title":"APIResponse[list[ReadChallengeSolution]]"},"APIResponse_list_ReadChallenge__":{"properties":{"success":{"type":"boolean","title":"Success"},"data":{"anyOf":[{"items":{"$ref":"#/components/schemas/ReadChallenge"},"type":"array"},{"type":"null"}],"title":"Data"},"error":{"anyOf":[{"$ref":"#/components/schemas/ErrorDetail"},{"type":"null"}]}},"type":"object","required":["success"],"title":"APIResponse[list[ReadChallenge]]"},"APIResponse_list_Tag__":{"properties":{"success":{"type":"boolean","title":"Success"},"data":{"anyOf":[{"items":{"$ref":"#/components/schemas/Tag"},"type":"array"},{"type":"null"}],"title":"Data"},"error":{"anyOf":[{"$ref":"#/components/schemas/ErrorDetail"},{"type":"null"}]}},"type":"object","required":["success"],"title":"APIResponse[list[Tag]]"},"Challenge":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"name":{"type":"string","maxLength":100,"title":"Name"},"is_free":{"type":"boolean","title":"Is Free"},"type":{"type":"string","title":"Type"},"difficulty":{"type":"integer","maximum":10.0,"exclusiveMinimum":0.0,"title":"Difficulty"},"description":{"type":"string","title":"Description"},"image_url":{"type":"string","title":"Image Url"},"created_at":{"type":"string","format":"date-time","title":"Created At","default":"2026-05-13T03:21:17.984859"},"last_edited":{"type":"string","format":"date-time","title":"Last Edited"}},"type":"object","required":["name","is_free","type","difficulty","description","image_url"],"title":"Challenge"},"ChallengeBody":{"properties":{"name":{"type":"string","maxLength":100,"title":"Name"},"is_free":{"type":"boolean","title":"Is Free"},"type":{"$ref":"#/components/schemas/ChallengeType"},"difficulty":{"type":"integer","maximum":10.0,"exclusiveMinimum":0.0,"title":"Difficulty"},"description":{"type":"string","title":"Description"},"image_url":{"type":"string","title":"Image Url"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags"}},"type":"object","required":["name","is_free","type","difficulty","description","image_url","tags"],"title":"ChallengeBody"},"ChallengeSolutionBody":{"properties":{"general_description":{"type":"string","title":"General Description"},"trouble_description":{"type":"string","title":"Trouble Description"},"total_rate":{"type":"integer","maximum":10.0,"minimum":1.0,"title":"Total Rate"},"total_difficulty":{"type":"integer","maximum":10.0,"minimum":1.0,"title":"Total Difficulty"}},"type":"object","required":["general_description","trouble_description","total_rate","total_difficulty"],"title":"ChallengeSolutionBody"},"ChallengeTagsLink":{"properties":{"challenge_id":{"type":"integer","title":"Challenge Id"},"tag_id":{"type":"integer","title":"Tag Id"}},"type":"object","required":["challenge_id","tag_id"],"title":"ChallengeTagsLink"},"ChallengeType":{"type":"string","enum":["Frontend","Backend","Fullstack"],"title":"ChallengeType"},"ErrorDetail":{"properties":{"code":{"type":"string","title":"Code"},"message":{"type":"string","title":"Message"}},"type":"object","required":["code","message"],"title":"ErrorDetail","description":"Standard error detail format"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ReadChallenge":{"properties":{"name":{"type":"string","maxLength":100,"title":"Name"},"is_free":{"type":"boolean","title":"Is Free"},"type":{"$ref":"#/components/schemas/ChallengeType"},"difficulty":{"type":"integer","maximum":10.0,"exclusiveMinimum":0.0,"title":"Difficulty"},"description":{"type":"string","title":"Description"},"image_url":{"type":"string","title":"Image Url"},"tags":{"items":{"$ref":"#/components/schemas/Tag"},"type":"array","title":"Tags"}},"type":"object","required":["name","is_free","type","difficulty","description","image_url","tags"],"title":"ReadChallenge"},"ReadChallengeSolution":{"properties":{"general_description":{"type":"string","title":"General Description"},"trouble_description":{"type":"string","title":"Trouble Description"},"total_rate":{"type":"integer","maximum":10.0,"minimum":1.0,"title":"Total Rate"},"total_difficulty":{"type":"integer","maximum":10.0,"minimum":1.0,"title":"Total Difficulty"},"id":{"type":"integer","title":"Id"},"challenge_id":{"type":"integer","title":"Challenge Id"},"user_id":{"type":"integer","title":"User Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"user":{"$ref":"#/components/schemas/ReadUser"}},"type":"object","required":["general_description","trouble_description","total_rate","total_difficulty","id","challenge_id","user_id","created_at","user"],"title":"ReadChallengeSolution"},"ReadUser":{"properties":{"id":{"type":"integer","title":"Id"},"github_id":{"type":"integer","title":"Github Id"},"username":{"type":"string","title":"Username"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"avatar_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Avatar Url"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","github_id","username","created_at"],"title":"ReadUser"},"Tag":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"name":{"type":"string","maxLength":100,"title":"Name"}},"type":"object","required":["name"],"title":"Tag"},"TagBody":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"TagBody"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}